Update Snakefile to include rules for fasterq-dump download, and merging technical replicates
Update pipeline. This is an update to the pipeline that would add fasterq-dump and merging technical replicates. The format of the input .tsv file also changed. The goal of changing the input was to make sure that the fasterq-dump could use the SRX ID to download the files. The files should be verified that they are download. You can also skip fasterq-dump if you download the files to the correct directory structure. The columns are BIO_REP (srx) and TECH_REP (srr).